requirements.txt
aiohttp>=3.9
python-dotenv>=1.0
gspread>=6.0
oauth2client>=4.1
apscheduler>=3.10
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
redis>=5.0.0
jinja2>=3.1.0
matplotlib>=3.7.0
plotly>=5.15.0
pandas>=2.0.0
python-telegram-bot>=20.0
passlib[bcrypt]>=1.7.4
bcrypt==4.3.0
python-jose[cryptography]>=3.3.0
